saree_gui/requirements.txt
requests
beautifulsoup4
transformers
torch
pillow
tk
